+++ neurons/text/prompting/miners/huggingface/koala/neuron.py
@@ -20,38 +20,35 @@ import bittensor
 from typing import List, Dict
 from transformers import AutoTokenizer, AutoModelForCausalLM
 
-from base import HuggingFaceMiner
-
-class NeoxtMiner( HuggingFaceMiner ):
-    arg_prefix: str = 'neoxt'
-    assistant_label: str = '<bot>:'
-    user_label: str = '<human>:'
+class KoalaMiner( bittensor.HuggingFaceMiner ):
+    arg_prefix: str = 'koala'
+    assistant_label: str = 'GPT:'
+    user_label: str = 'USER:'
     system_label: str = ''
 
     def load_tokenizer( self ):
-        return AutoTokenizer.from_pretrained( self.config.neoxt.model_name )
+        return AutoTokenizer.from_pretrained( self.config.koala.model_name, use_fast=False )
 
     def load_model( self ):
-        return AutoModelForCausalLM.from_pretrained( self.config.neoxt.model_name, torch_dtype = torch.float16, low_cpu_mem_usage=True )
+        return AutoModelForCausalLM.from_pretrained( self.config.koala.model_name, torch_dtype = torch.float16, low_cpu_mem_usage=True )
 
     def forward( self, messages: List[Dict[str, str]] ) -> str:
-        history = self.process_history( messages )
-        prompt = history + self.assistant_label
-        input_ids = self.tokenizer.encode( prompt, return_tensors="pt" ).to( self.config.neoxt.device )
+        history = self._process_history( messages )
+        prompt = history + self.system_label
+        input_ids = self.tokenizer.encode( prompt, return_tensors="pt" ).to( self.config.koala.device )
         output = self.model.generate(
             input_ids,
-            max_length=input_ids.shape[1] + self.config.neoxt.max_new_tokens,
-            temperature=self.config.neoxt.temperature,
-            do_sample=self.config.neoxt.do_sample,
+            max_length=input_ids.shape[1] + self.config.koala.max_new_tokens,
+            temperature=self.config.koala.temperature,
+            do_sample=self.config.koala.do_sample,
             pad_token_id=self.tokenizer.eos_token_id,
         )
-        generated_text = self.tokenizer.decode( output[0][input_ids.shape[1]:], skip_special_tokens=True )
-        generation = generated_text.split( "<human>" )[0].strip()
+        generation = self.tokenizer.decode( output[0][input_ids.shape[1]:], skip_special_tokens=True )
 
-        bittensor.logging.debug( "Message: " + str( messages ).replace( "<","-" ).replace( ">","-" ) )
-        bittensor.logging.debug( "Generation: " + str( generation ).replace( "<","-" ).replace( ">","-" ) )
+        bittensor.logging.debug( "Message: " + str( messages ) )
+        bittensor.logging.debug( "Generation: " + str( generation ) )
         return generation
 
 if __name__ == "__main__":
     bittensor.utils.version_checking()
-    NeoxtMiner().run()
+    KoalaMiner().run()
