+++ neurons/text/prompting/miners/huggingface/stabilityai/stabilityai_requirements.txt
