+++ neurons/text/prompting/miners/huggingface/vicuna/neuron.py
@@ -1,5 +1,5 @@
 # The MIT License (MIT)
-# Copyright © 2023 Opentensor Foundation
+# Copyright © 2023 Yuma Rao
 
 # Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
 # documentation files (the “Software”), to deal in the Software without restriction, including without limitation
@@ -20,37 +20,45 @@ import bittensor
 from typing import List, Dict
 from transformers import AutoTokenizer, AutoModelForCausalLM
 
-from base import HuggingFaceMiner
+class VicunaMiner( bittensor.HuggingFaceMiner ):
 
-class WizardVicunaMiner( HuggingFaceMiner ):
-    arg_prefix = "wiz_vic"
-    system_label = ""
-    assistant_label = "ASSISTANT:"
-    user_label = "USER:"
+    arg_prefix: str = 'vicuna'
+    system_label: str = ''
+    assistant_label: str = 'ASSISTANT:'
+    user_label: str = 'USER:'
+
+    def __init__( self ):
+        super( VicunaMiner, self ).__init__()
+        print ( self.config )
 
     def load_tokenizer( self ):
-        return AutoTokenizer.from_pretrained( self.config.wiz_vic.model_name, use_fast=False )
+        return AutoTokenizer.from_pretrained( self.config.vicuna.model_name, use_fast=False )
 
     def load_model( self ):
-        return AutoModelForCausalLM.from_pretrained( self.config.wiz_vic.model_name, torch_dtype=torch.float16, low_cpu_mem_usage=True )
+        return AutoModelForCausalLM.from_pretrained( self.config.vicuna.model_name, torch_dtype = torch.float16, low_cpu_mem_usage=True )
 
-    def forward( self, messages: List[Dict[str, str]] ) -> str:
+    def forward(self, messages: List[Dict[str, str]]) -> str:
+
         history = self.process_history( messages )
         prompt = history + self.assistant_label
-        input_ids = self.tokenizer.encode( prompt, return_tensors="pt" ).to( self.config.wiz_vic.device )
+        print(prompt)
+
+        input_ids = self.tokenizer.encode( prompt, return_tensors="pt" ).to( self.config.vicuna.device )
         output = self.model.generate(
             input_ids,
-            max_length=input_ids.shape[1] + self.config.wiz_vic.max_new_tokens,
-            temperature=self.config.wiz_vic.temperature,
-            do_sample=self.config.wiz_vic.do_sample,
+            max_length=input_ids.shape[1] + self.config.vicuna.max_new_tokens,
+            temperature=self.config.vicuna.temperature,
+            do_sample=self.config.vicuna.do_sample,
             pad_token_id=self.tokenizer.eos_token_id,
         )
+
         generation = self.tokenizer.decode( output[0][input_ids.shape[1]:], skip_special_tokens=True )
+        print(generation)
 
         bittensor.logging.debug( "Message: " + str( messages ) )
-        bittensor.logging.debug( "Generation: " + str( generation) )
+        bittensor.logging.debug( "Generation: " + str( generation ) )
         return generation
 
 if __name__ == "__main__":
     bittensor.utils.version_checking()
-    WizardVicunaMiner().run()
+    VicunaMiner().run()
