+++ neurons/text/prompting/miners/huggingface/wizard_vicuna/wizardvicuna_miner.py
@@ -1,5 +1,5 @@
 # The MIT License (MIT)
-# Copyright © 2023 Yuma Rao
+# Copyright © 2023 Opentensor Foundation
 
 # Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
 # documentation files (the “Software”), to deal in the Software without restriction, including without limitation
@@ -20,39 +20,35 @@ import bittensor
 from typing import List, Dict
 from transformers import AutoTokenizer, AutoModelForCausalLM
 
-from base import HuggingFaceMiner
-
-class PythiaMiner( HuggingFaceMiner ):
-
-    arg_prefix: str = 'pythia'
-    assistant_label: str = '<bot>:'
-    user_label: str = '<human>:'
-    system_label: str = ''
+class WizardVicunaMiner( bittensor.HuggingFaceMiner ):
+    arg_prefix = "wiz_vic"
+    system_label = ""
+    assistant_label = "ASSISTANT:"
+    user_label = "USER:"
 
     def load_tokenizer( self ):
-        return AutoTokenizer.from_pretrained( self.config.pythia.model_name )
+        return AutoTokenizer.from_pretrained( self.config.wiz_vic.model_name, use_fast=False )
 
     def load_model( self ):
-        return AutoModelForCausalLM.from_pretrained( self.config.pythia.model_name, torch_dtype = torch.float16, low_cpu_mem_usage=True )
-    
-    def forward(self, messages: List[Dict[str, str]]) -> str:
+        return AutoModelForCausalLM.from_pretrained( self.config.wiz_vic.model_name, torch_dtype=torch.float16, low_cpu_mem_usage=True )
+
+    def forward( self, messages: List[Dict[str, str]] ) -> str:
         history = self.process_history( messages )
         prompt = history + self.assistant_label
-        input_ids = self.tokenizer.encode( prompt, return_tensors="pt" ).to( self.config.pythia.device )
+        input_ids = self.tokenizer.encode( prompt, return_tensors="pt" ).to( self.config.wiz_vic.device )
         output = self.model.generate(
             input_ids,
-            max_length=input_ids.shape[1] + self.config.pythia.max_new_tokens,
-            temperature=self.config.pythia.temperature,
-            do_sample=self.config.pythia.do_sample,
+            max_length=input_ids.shape[1] + self.config.wiz_vic.max_new_tokens,
+            temperature=self.config.wiz_vic.temperature,
+            do_sample=self.config.wiz_vic.do_sample,
             pad_token_id=self.tokenizer.eos_token_id,
         )
-        generated_text = self.tokenizer.decode( output[0][input_ids.shape[1]:], skip_special_tokens=True )
-        generation = generated_text.split( "<human>" )[0].strip()
+        generation = self.tokenizer.decode( output[0][input_ids.shape[1]:], skip_special_tokens=True )
 
-        bittensor.logging.debug("Message: " + str( messages ).replace( "<","-" ).replace( ">","-" ) )
-        bittensor.logging.debug("Generation: " + str( generation ).replace( "<","-" ).replace( ">","-" ) )
+        bittensor.logging.debug( "Message: " + str( messages ) )
+        bittensor.logging.debug( "Generation: " + str( generation) )
         return generation
 
 if __name__ == "__main__":
     bittensor.utils.version_checking()
-    PythiaMiner().run()
+    WizardVicunaMiner().run()
